Refactoring and cleanup for SPSP and ILQP
