+++ src/main/java/org/interledger/ilp/ledger/adaptor/rest/RestLedgerAdaptor.java
@@ -9,14 +9,14 @@ import java.util.Set;
 import java.util.UUID;
 
 import org.interledger.cryptoconditions.Fulfillment;
-import org.interledger.ilp.core.InterledgerAddress;
-import org.interledger.ilp.core.ledger.LedgerAdaptor;
-import org.interledger.ilp.core.ledger.events.LedgerEventHandler;
-import org.interledger.ilp.core.ledger.model.AccountInfo;
-import org.interledger.ilp.core.ledger.model.LedgerInfo;
-import org.interledger.ilp.core.ledger.model.LedgerMessage;
-import org.interledger.ilp.core.ledger.model.LedgerTransfer;
-import org.interledger.ilp.core.ledger.model.TransferRejectedReason;
+import org.interledger.ilp.InterledgerAddress;
+import org.interledger.ilp.ledger.LedgerAdaptor;
+import org.interledger.ilp.ledger.events.LedgerEventHandler;
+import org.interledger.ilp.ledger.model.AccountInfo;
+import org.interledger.ilp.ledger.model.LedgerInfo;
+import org.interledger.ilp.ledger.model.LedgerMessage;
+import org.interledger.ilp.ledger.model.LedgerTransfer;
+import org.interledger.ilp.ledger.model.TransferRejectedReason;
 import org.interledger.ilp.ledger.adaptor.rest.exceptions.RestServiceException;
 import org.interledger.ilp.ledger.adaptor.rest.service.RestLedgerAccountService;
 import org.interledger.ilp.ledger.adaptor.rest.service.RestLedgerAuthTokenService;
