+++ src/main/java/org/interledger/ilp/ledger/adaptor/rest/service/RestLedgerJsonConverter.java
@@ -21,16 +21,12 @@ import javax.money.format.MonetaryAmountFormat;
 
 import org.interledger.cryptoconditions.uri.CryptoConditionUri;
 import org.interledger.cryptoconditions.uri.URIEncodingException;
+import org.interledger.ilp.InterledgerAddress;
 import org.interledger.ilp.client.exceptions.DataModelTranslationException;
 import org.interledger.ilp.client.model.ClientAccountInfo;
 import org.interledger.ilp.client.model.ClientLedgerInfo;
 import org.interledger.ilp.client.model.ClientLedgerMessage;
 import org.interledger.ilp.client.model.ClientLedgerTransfer;
-import org.interledger.ilp.core.InterledgerAddress;
-import org.interledger.ilp.core.ledger.model.AccountInfo;
-import org.interledger.ilp.core.ledger.model.LedgerInfo;
-import org.interledger.ilp.core.ledger.model.LedgerMessage;
-import org.interledger.ilp.core.ledger.model.LedgerTransfer;
 import org.interledger.ilp.ledger.adaptor.rest.ServiceUrls;
 import org.interledger.ilp.ledger.adaptor.rest.json.JsonAccountInfo;
 import org.interledger.ilp.ledger.adaptor.rest.json.JsonLedgerInfo;
@@ -38,16 +34,20 @@ import org.interledger.ilp.ledger.adaptor.rest.json.JsonLedgerMessage;
 import org.interledger.ilp.ledger.adaptor.rest.json.JsonLedgerTransfer;
 import org.interledger.ilp.ledger.adaptor.rest.json.JsonLedgerTransferAccountEntry;
 import org.interledger.ilp.ledger.adaptor.rest.json.JsonValidator;
-import org.interledger.ilp.ledger.adaptor.rest.json.LedgerSpecificDecimalMonetaryAmountFormat;
+import org.interledger.ilp.ledger.model.AccountInfo;
+import org.interledger.ilp.ledger.model.LedgerInfo;
+import org.interledger.ilp.ledger.model.LedgerMessage;
+import org.interledger.ilp.ledger.model.LedgerTransfer;
+import org.interledger.ilp.ledger.money.format.LedgerSpecificDecimalMonetaryAmountFormat;
 import org.interledger.ilqp.client.model.ClientQuoteErrorResponse;
 import org.interledger.ilqp.client.model.ClientQuoteResponse;
-import org.interledger.ilqp.core.model.QuoteRequest;
 import org.interledger.ilqp.json.JsonErrorResponse;
 import org.interledger.ilqp.json.JsonErrorResponseEnvelope;
 import org.interledger.ilqp.json.JsonQuoteRequest;
 import org.interledger.ilqp.json.JsonQuoteRequestEnvelope;
 import org.interledger.ilqp.json.JsonQuoteResponse;
 import org.interledger.ilqp.json.JsonQuoteResponseEnvelope;
+import org.interledger.quoting.model.QuoteRequest;
 import org.springframework.web.util.UriTemplate;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
