+++ src/main/java/org/interledger/ilp/ledger/adaptor/rest/service/RestLedgerMetaService.java
@@ -6,7 +6,7 @@ import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
 
-import org.interledger.ilp.core.ledger.model.LedgerInfo;
+import org.interledger.ilp.ledger.model.LedgerInfo;
 import org.interledger.ilp.ledger.adaptor.rest.ServiceUrls;
 import org.interledger.ilp.ledger.adaptor.rest.json.JsonConnectorInfo;
 import org.interledger.ilp.ledger.adaptor.rest.json.JsonLedgerInfo;
