+++ src/main/java/org/interledger/ilp/ledger/adaptor/ws/JsonRpcLedgerWebSocketChannel.java
@@ -6,7 +6,7 @@ import org.interledger.ilp.client.events.ClientLedgerConnectEvent;
 import org.interledger.ilp.client.events.ClientLedgerErrorEvent;
 import org.interledger.ilp.client.events.ClientLedgerMessageEvent;
 import org.interledger.ilp.client.events.ClientLedgerTransferEvent;
-import org.interledger.ilp.core.ledger.events.LedgerEventHandler;
+import org.interledger.ilp.ledger.events.LedgerEventHandler;
 import org.interledger.ilp.ledger.adaptor.rest.json.JsonLedgerMessage;
 import org.interledger.ilp.ledger.adaptor.rest.json.JsonLedgerTransfer;
 import org.interledger.ilp.ledger.adaptor.rest.service.RestLedgerJsonConverter;
