+++ src/main/java/org/interledger/ilqp/json/JsonQuoteRequestEnvelope.java
@@ -1,6 +1,6 @@
 package org.interledger.ilqp.json;
 
-import org.interledger.ilp.core.ledger.model.MessageData;
+import org.interledger.ilp.ledger.model.MessageData;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
